--- a/stdlib/public/SwiftShims/GlobalObjects.h
+++ b/stdlib/public/SwiftShims/GlobalObjects.h
@@ -45,6 +45,13 @@ struct _SwiftUnsafeBitMap {
   __swift_intptr_t bitCount;
 };
 
+struct _SwiftHashTable {
+  __swift_intptr_t scale;
+  __swift_intptr_t capacity;
+  __swift_intptr_t count;
+  void *rawMap;
+};
+
 struct _SwiftDictionaryBodyStorage {
   __swift_intptr_t capacity;
   __swift_intptr_t count;
@@ -54,10 +61,10 @@ struct _SwiftDictionaryBodyStorage {
 };
 
 struct _SwiftSetBodyStorage {
-  __swift_intptr_t capacity;
-  __swift_intptr_t count;
-  struct _SwiftUnsafeBitMap initializedEntries;
-  void *keys;
+  struct _SwiftHashTable hashTable;
+  __swift_uint64_t seed0;
+  __swift_uint64_t seed1;
+  void *rawElements;
 };
 
 struct _SwiftEmptyDictionaryStorage {