--- a/stdlib/public/SwiftShims/GlobalObjects.h
+++ b/stdlib/public/SwiftShims/GlobalObjects.h
@@ -46,8 +46,8 @@ struct _SwiftUnsafeBitMap {
 };
 
 struct _SwiftHashTable {
-  __swift_intptr_t capacity;
   __swift_intptr_t count;
+  __swift_intptr_t capacity;
   __swift_intptr_t scale;
   void *map;
 };