… stdlib/public/core/CMakeLists.txt
… stdlib/public/core/GroupInfo.json
… stdlib/public/core/Hasher.swift
… stdlib/public/core/Hashing.swift
… stdlib/public/stubs/GlobalObjects.cpp
… validation-test/stdlib/Bitmap.swift
… validation-test/stdlib/Set.swift