- Extract low-level hash table operations into a type-agnostic `_HashTable` struct, parts of which can potentially be made resilient.
- Store 7 additional bits of the hash value directly in each entry's metadata field. (The hope is that this will allow us to use a much higher max load factor, by making collision chains dramatically cheaper.)
- Updated storage class hierarchy and internal storage layout for `Set` and (soon) `Dictionary`.
- New representation for the type-punned empty singleton storage.
- Slightly less memory overhead for deferred bridging. (This is especially relevant for `Dictionary` where only one of the element types may need non-verbatim bridging.)

This hasn't been fully debugged yet, and also needs to be extensively benchmarked/fine-tuned.

An lldb data formatter update will need to land before merging this PR.