@lorentey 2018-08-21
@swift-ci benchmark