@lorentey 2018-08-22
@swift-ci please benchmark