@lorentey 2018-08-22
TEST                                                      | OLD    | NEW    | DELTA  | SPEEDUP   
---                                                       | ---    | ---    | ---    | ---       
SetIntersect                                              | 587    | 1022   | +74.1% | **0.57x** 
SetUnion                                                  | 4148   | 6054   | +45.9% | **0.69x** 

That's funny! These results are the opposite of my own `Set<Int>` benchmarking.