@lorentey 2018-08-23
![set - 4 tasks](https://user-images.githubusercontent.com/608696/44469358-3a78ea80-a61f-11e8-9be7-ccae77ffff6e.png)

Separately benchmarking SetIntersect in detail reveals an important clue -- the benchmark as it currently stands measures the intersection between two non-overlapping sets. I can't locally reproduce the +75% slowdown -- this use-case benchmarks roughly the same before/after this PR.

Fixing the benchmark to include a non-empty intersection (with ~25% of elements in both sets) shows clear improvements over the non-overlapping case. (See `SetIntersection2` above.) This is rather surprising, as the algorithm actually needs to do *more* work in the overlapping case, not less.