@lorentey 2018-08-23
`SetIntersection` uses a random number generator, which generates duplicate elements in the `SetIntersection2` case, making `SetIntersection` and `SetIntersection2` not directly comparable. I submitted #18928 to fix these benchmarks on master.

Updated chart:
![set - 4 tasks](https://user-images.githubusercontent.com/608696/44549421-ee11d580-a718-11e8-8cee-8003cc5f247a.png)

Note how the new algorithm is faster at almost every stage. It gets relatively faster as the % of common elements increase. The intersection of two equal sets is calculated particularly efficiently:

![set - 4 tasks](https://user-images.githubusercontent.com/608696/44549596-7b552a00-a719-11e8-9238-beb66bd4c7cc.png)

The new algorithm detects that there is no need to construct a third set, making it even faster than the 0% case for many sizes. (This is because looking up an existing element is faster than one that's not in the hash table.)