@lorentey 2018-08-25
@swift-ci please test linux platform