@lorentey 2018-08-25
Subtraction optimizations in the latest updates should help clean up benchmark regressions.

@swift-ci please smoke benchmark staging