@lorentey 2018-08-28
Latest commits get us to master's performance or better for most benchmarks, at the cost of disabling hash table layout resilience. 

These commits have not yet shown up in this PR; let's see if CI picks them up anyway.

@swift-ci please smoke benchmark staging