@lorentey 2018-08-28
I see GitHub has caught up now.

@swift-ci please smoke benchmark staging