@lorentey 2018-08-28
@swift-ci please benchmark