@lorentey 2018-08-30
@swift-ci please smoke benchmark staging